refactor customer logic into customercontroller
